refactor: better md management
