+++ core-md/src/androidMain/kotlin/com/github/diegoberaldin/raccoonforlemmy/core/markdown/provider/MarkwonProvider.kt
@@ -1,4 +1,4 @@
-package com.github.diegoberaldin.raccoonforlemmy.core.markdown.compose
+package com.github.diegoberaldin.raccoonforlemmy.core.markdown.provider
 
 import io.noties.markwon.Markwon
